Extracted render component by flag props from CounterDisplay
