+++ components/Counter/CounterDisplay.jsx
@@ -1,11 +1,11 @@
 import React, { useState } from 'react'
-import { Text, View, StyleSheet, TextInput } from 'react-native'
+import { View, StyleSheet } from 'react-native'
 import { PanGestureHandler, TapGestureHandler } from 'react-native-gesture-handler'
 import Animated, { useSharedValue } from 'react-native-reanimated'
 import { animatedMove } from '../../utils/animatedStyles'
 import { responsivePixel } from '../../utils/responsivePixel'
 import { counterDisplayHandleGesture } from '../../utils/counterDisplayHandleGesture'
-import { updateCounter } from '../../services/updateCounter'
+import { display } from '../../render-props/displayFlagRender'
 
 const CounterDisplay = ({ id, counterName, setCount, count }) => {
   // Center of the Counter
@@ -26,28 +26,7 @@ const CounterDisplay = ({ id, counterName, setCount, count }) => {
   const handleDoubleTap = () => {
     setDoubleTapped(true)
   }
-  const handleSubmiting = () => {
-    setDoubleTapped(false)
-    let trimedVal
-    // Verifiying data and cleaning the string from empty spaces
 
-    if (count === undefined) {
-      trimedVal = ''
-    } else {
-      trimedVal = count.trim()
-    }
-
-    // Regular expresion that takes out all non digit characters and leadding zeros
-    let filteredVal = trimedVal.replace(/^0+/, "").replace(/\D/g, "")
-    if (filteredVal === '') filteredVal = "0"
-    setCount(JSON.parse(filteredVal))
-    updateCounter(id, counterName, filteredVal)
-  }
-  const display = () => {
-    return doubleTapped
-      ? <TextInput style={[styles.counterDisplay, { backgroundColor: backgroundColor }]} onChangeText={setCount} onSubmitEditing={handleSubmiting}>{count}</TextInput>
-      : <Text style={[styles.counterDisplay, { backgroundColor: backgroundColor }]}>{count}</Text>
-  }
   // Event hanlder provided by Reanimated that is specifically designed to work with the gesture-handler package
   const eventHandler = counterDisplayHandleGesture(x, y, startingPosition, id, counterName, setCount, count, setReseting)
   return <View>
@@ -56,7 +35,7 @@ const CounterDisplay = ({ id, counterName, setCount, count }) => {
         <TapGestureHandler
           numberOfTaps={2}
           onActivated={handleDoubleTap}>
-          {display()}
+          {display(id, counterName, doubleTapped, setDoubleTapped, count, setCount, backgroundColor)}
         </TapGestureHandler>
       </Animated.View>
     </PanGestureHandler>
@@ -85,21 +64,6 @@ const styles = StyleSheet.create({
 
     elevation: 2,
   },
-  counterDisplay: {
-    textAlign: 'center',
-    fontSize: responsivePixel(45),
-    minWidth: responsivePixel(110),
-    borderRadius: 40,
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-
-    elevation: 5,
-  }
 })
 
 export default CounterDisplay
